Move test to correct context
